replace placekitten with placebear
unfortunately placekitten.com currently doesn't work which breaks the
example code appearance. We should replace these links even though this
isn't code directly related to the exercises because students might be
trying to follow along (or use this service when trying to complete the
example).

I chose placebear.com because it imitates the same API and
bears deserve a little more love.